Smart global governance (#701)
* create folder and apiDefinitionSwagger

* Add files via upload

Initial delivery

* Delete icon.png

On demand of Microsoft. Not needed for open sourcing.

* Delete settings.json

On demand of Microsoft. Not needed for open sourcing.

* Add files via upload

Title of the connector changed to "Smart Global Governance".
Description expanded to highlight the functionality and value that the connector offers.
Contact information changed to "Smart Global Governance Support Team".

* Add files via upload

Added new responses according to HTTP codes.

* Add files via upload

* Add files via upload

Changed a definition file reference that was wrong.
Changed "compliant" property to "complaint".

* Add files via upload

Did the reformating you expected.

* Add files via upload

Reindented the content to be compliant with your expectations.

Co-authored-by: Philippe-Tivolle <77966693+Philippe-Tivolle@users.noreply.github.com>